move tests to Common modules
