+++ integration-tests/oauth2-servlet/src/test/resources/testRunner.yml
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-implementation: com.okta.spring.tests.oauth2.SpringApplicationUnderTest
+implementation: com.okta.spring.tests.common.servlet.SpringApplicationUnderTest
 scenarios:
 
   code-flow-remote-validation:
