+++ pom.xml
@@ -105,8 +105,24 @@
             </dependency>
             <dependency>
                 <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-common-servlet</artifactId>
+                <version>2.0.2-SNAPSHOT</version>
+                <classifier>tests</classifier>
+                <type>test-jar</type>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
                 <artifactId>okta-spring-boot-integration-tests-common-reactive</artifactId>
                 <version>2.0.2-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>com.okta.spring</groupId>
+                <artifactId>okta-spring-boot-integration-tests-common-reactive</artifactId>
+                <version>2.0.2-SNAPSHOT</version>
+                <classifier>tests</classifier>
+                <type>test-jar</type>
+                <scope>test</scope>
             </dependency>
 
             <dependency>
